+++ src/codec/CommandsCodec.ts
@@ -5,7 +5,6 @@ import {Commands} from "../model/Commands";
 import {Commands as PbCommands} from "../generated/com/digitalasset/ledger/api/v1/commands_pb";
 import {Codec} from "./Codec";
 import {CommandCodec} from "./CommandCodec";
-import {TimestampCodec} from "./TimestampCodec";
 
 export const CommandsCodec: Codec<PbCommands, Commands> = {
     deserialize(commands: PbCommands): Commands {
@@ -13,8 +12,6 @@ export const CommandsCodec: Codec<PbCommands, Commands> = {
             applicationId: commands.getApplicationId(),
             commandId: commands.getCommandId(),
             party: commands.getParty(),
-            ledgerEffectiveTime: TimestampCodec.deserialize(commands.getLedgerEffectiveTime()!),
-            maximumRecordTime: TimestampCodec.deserialize(commands.getMaximumRecordTime()!),
             list: commands.getCommandsList().map((command) => CommandCodec.deserialize(command))
         };
         const workflowId = commands.getWorkflowId();
@@ -27,8 +24,6 @@ export const CommandsCodec: Codec<PbCommands, Commands> = {
         const result = new PbCommands();
         result.setCommandId(commands.commandId);
         result.setParty(commands.party);
-        result.setLedgerEffectiveTime(TimestampCodec.serialize(commands.ledgerEffectiveTime));
-        result.setMaximumRecordTime(TimestampCodec.serialize(commands.maximumRecordTime));
         result.setCommandsList(commands.list.map((command) => CommandCodec.serialize(command)));
         if (commands.workflowId) {
             result.setWorkflowId(commands.workflowId);